Fixed bugs with fakes, latches and cell voltage faking. Some ifdef stuff too

+++ firmware/quintuna/BMS/test/test_fakes.hpp
@@ -42,7 +42,7 @@ namespace imd
 namespace segments
 {
     void setCellVoltages(const std::array<std::array<float, CELLS_PER_SEGMENT>, NUM_SEGMENTS> &voltages);
-    void setCellVoltage(size_t segment, size_t cell, uint16_t voltage);
+    void setCellVoltage(size_t segment, size_t cell, float voltage);
     void setPackVoltageEvenly(float pack_voltage);
     void setCellTemperatures(const std::array<std::array<float, AUX_REGS_PER_SEGMENT>, NUM_SEGMENTS> &temperatures);
